requirements.txt
numpy
google-cloud-firestore
google-auth
Flask
flask-cors
Pillow
